师资指数分页后端ok

+++ src/main/java/org/wlgzs/index_evaluation/controller/TeachersStructureController.java
@@ -1,10 +1,13 @@
 package org.wlgzs.index_evaluation.controller;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
@@ -18,6 +21,7 @@ import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.io.InputStream;
 import java.text.DecimalFormat;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -37,41 +41,87 @@ public class TeachersStructureController {
     private YearService yearService;
 
     @GetMapping("/delete")
-    public ModelAndView delete(Integer year){
+    public ModelAndView delete(Integer year,
+                               @RequestParam(name = "pageNum", defaultValue = "1") int pageNum,
+                               @RequestParam(name = "pageSize", defaultValue = "16") int pageSize){
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.setViewName("result");
         List<Year> allYear = yearService.findAllYear();
         modelAndView.addObject("allYear",allYear);
         List<TeachersStructure> byYear = teachersStructureService.findByYear(year);
-        List<TeachersStructure> allTeachersStructure = teachersStructureService.findAll();
-        modelAndView.addObject("allTeachersStructure",allTeachersStructure);
+        List<TeachersStructure> allTeachersStructure;
+        if (year == null){
+            modelAndView.addObject("msg","请选择年份后重试！");
+            Page<TeachersStructure> practiceQueryWrapper = new Page<>(pageNum,pageSize);
+            QueryWrapper<TeachersStructure> queryWrapper = new QueryWrapper<>();
+            IPage<TeachersStructure> iPage = teachersStructureService.page(practiceQueryWrapper,queryWrapper);
+            modelAndView.addObject("current",iPage.getCurrent());//当前页数
+            modelAndView.addObject("pages",iPage.getPages());//总页数
+            modelAndView.addObject("allTeachersStructure",iPage.getRecords());//所有的数据集合
+            return modelAndView;
+        }
         for (TeachersStructure t :byYear){
             if (teachersStructureService.delete(t) == 0){
                 modelAndView.addObject("msg","删除出错，请重试！");
+                Page<TeachersStructure> practiceQueryWrapper = new Page<>(pageNum,pageSize);
+                QueryWrapper<TeachersStructure> queryWrapper = new QueryWrapper<>();
+                IPage<TeachersStructure> iPage = teachersStructureService.page(practiceQueryWrapper,queryWrapper);
+                modelAndView.addObject("current",iPage.getCurrent());//当前页数
+                modelAndView.addObject("pages",iPage.getPages());//总页数
+                modelAndView.addObject("allTeachersStructure",iPage.getRecords());//所有的数据集合
                 return modelAndView;
             }
         }
         modelAndView.addObject("msg","删除成功");
+        Page<TeachersStructure> practiceQueryWrapper = new Page<>(pageNum,pageSize);
+        QueryWrapper<TeachersStructure> queryWrapper = new QueryWrapper<>();
+        IPage<TeachersStructure> iPage = teachersStructureService.page(practiceQueryWrapper,queryWrapper);
+        modelAndView.addObject("current",iPage.getCurrent());//当前页数
+        modelAndView.addObject("pages",iPage.getPages());//总页数
+        modelAndView.addObject("allTeachersStructure",iPage.getRecords());//所有的数据集合
         return modelAndView;
     }
 
     @GetMapping("/to")
-    public ModelAndView to(){
+    public ModelAndView to(@RequestParam(name = "pageNum", defaultValue = "1") int pageNum,
+                           @RequestParam(name = "pageSize", defaultValue = "16") int pageSize){
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.setViewName("result");
         List<Year> allYear = yearService.findAllYear();
         modelAndView.addObject("allYear",allYear);
-        List<TeachersStructure> allTeachersStructure = teachersStructureService.findAll();
-        modelAndView.addObject("allTeachersStructure",allTeachersStructure);
+        Page<TeachersStructure> practiceQueryWrapper = new Page<>(pageNum,pageSize);
+        QueryWrapper<TeachersStructure> queryWrapper = new QueryWrapper<>();
+        IPage<TeachersStructure> iPage = teachersStructureService.page(practiceQueryWrapper,queryWrapper);
+        modelAndView.addObject("current",iPage.getCurrent());//当前页数
+        modelAndView.addObject("pages",iPage.getPages());//总页数
+        modelAndView.addObject("allTeachersStructure",iPage.getRecords());//所有的数据集合
         return modelAndView;
     }
 
     @RequestMapping("/import")
-    public void impor(HttpServletRequest request, Integer year) throws IOException {
+    public ModelAndView impor(HttpServletRequest request, Integer year,
+                              @RequestParam(name = "pageNum", defaultValue = "1") int pageNum,
+                              @RequestParam(name = "pageSize", defaultValue = "16") int pageSize) throws IOException {
+        ModelAndView modelAndView = new ModelAndView();
+        List<Year> allYear = yearService.findAllYear();
+        modelAndView.addObject("allYear",allYear);
+        List<TeachersStructure> allTeachersStructure;
+        modelAndView.setViewName("result");
         //获取上传的文件
         MultipartHttpServletRequest multipart = (MultipartHttpServletRequest) request;
         MultipartFile file = multipart.getFile("upfile");
+        log.info(file.getOriginalFilename().equals("师资结构指数样表.xlsx"));
         InputStream in = file.getInputStream();
+        if (!file.getOriginalFilename().equals("师资结构指数样表.xlsx")){
+            modelAndView.addObject("msg","上传文件错误，请确认是师资结构指数样表.xlsx！");
+            Page<TeachersStructure> practiceQueryWrapper = new Page<>(pageNum,pageSize);
+            QueryWrapper<TeachersStructure> queryWrapper = new QueryWrapper<>();
+            IPage<TeachersStructure> iPage = teachersStructureService.page(practiceQueryWrapper,queryWrapper);
+            modelAndView.addObject("current",iPage.getCurrent());//当前页数
+            modelAndView.addObject("pages",iPage.getPages());//总页数
+            modelAndView.addObject("allTeachersStructure",iPage.getRecords());//所有的数据集合
+            return modelAndView;
+        }
         List<TeachersStructure> teachersStructures = teachersStructureService.importExcelInfo(in, file);
         DecimalFormat df2 = new DecimalFormat("#.00");
         DecimalFormat df3 = new DecimalFormat("#.000");
@@ -129,8 +179,13 @@ public class TeachersStructureController {
             t.setA2(Double.parseDouble(df2.format((t.getM1()*33.7+t.getGraNum()/t.getTeaNum()/max0*32.9+t.getSenNum()/t.getTeaNum()/max1*33.9)*0.1107)));
             teachersStructureService.update(t);
         }
-
-
-
+        modelAndView.addObject("msg","上传文件成功！");
+        Page<TeachersStructure> practiceQueryWrapper = new Page<>(pageNum,pageSize);
+        QueryWrapper<TeachersStructure> queryWrapper = new QueryWrapper<>();
+        IPage<TeachersStructure> iPage = teachersStructureService.page(practiceQueryWrapper,queryWrapper);
+        modelAndView.addObject("current",iPage.getCurrent());//当前页数
+        modelAndView.addObject("pages",iPage.getPages());//总页数
+        modelAndView.addObject("allTeachersStructure",iPage.getRecords());//所有的数据集合
+        return modelAndView;
     }
 }
